Refactor general structure
