--- test/globals.d.ts
@@ -1,10 +1,0 @@
-import { Scroll } from "../src/scroll";
-
-declare global {
-    const Scroll: Scroll;
-    const divScroll: Scroll;
-    const scrollable: HTMLElement;
-    const element: HTMLElement;
-    const element1: HTMLElement
-}
-
