--- test/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-    <head>
-        <title>Testing</title>
-        <link rel="stylesheet" type="text/css", href="./local/index.css" />
-    </head>
-    <body>
-        <div id="scrollable">
-            <div id="element"> </div>
-        </div>
-        <div id="element1"> </div>
-    </body>
-    <script src="./loadScroll.ts"> </script>
-</html>
