--- tsconfig.build.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "baseUrl": ".",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "noImplicitAny": false,
-        "noImplicitThis": true,
-        "noUnusedLocals": true,
-        "outDir": "./dist/",
-        "rootDir": ".",
-        "declaration": true,
-        "target": "es6",
-        "paths": {
-            "*": [
-                "node_modules/*"
-            ]
-        }
-    },
-    "include": [
-        "./index.ts"
-    ]
-}
