+++ tsconfig.json
@@ -17,6 +17,6 @@
         }
     },
     "include": [
-        "./test"
+        "./index.ts"
     ]
 }
